# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ChannelCreate": {
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[^\\s#@/]{1,80}$",
            "title": "Name",
            "type": "string"
          },
          "purpose": {
            "anyOf": [
              {
                "maxLength": 250,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "topic": {
            "anyOf": [
              {
                "maxLength": 250,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "type": {
            "default": "public",
            "enum": [
              "public",
              "private"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ChannelCreate",
        "type": "object"
      },
...
      "ChannelOut": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "dm_user_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dm User Ids"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "last_seq": {
            "title": "Last Seq",
            "type": "integer"
          },
          "member_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Count"
          },
          "membership": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationPreferenceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "posting_policy": {
            "default": "everyone",
            "enum": [
              "everyone",
              "owners"
            ],
            "title": "Posting Policy",
            "type": "string"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "read_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadStateOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "times_owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Times Owner Id"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "type": {
            "enum": [
              "public",
              "private",
              "dm",
              "group_dm"
            ],
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "topic",
          "purpose",
          "archived",
          "created_by",
          "last_seq",
          "last_message_at",
          "created_at",
          "updated_at",
          "membership",
          "dm_user_ids"
        ],
        "title": "ChannelOut",
        "type": "object"
      },
      "ChannelReadStateOut": {
        "description": "One channel's read state after POST /channels/read-all (M12a).",
        "properties": {
          "channel_id": {
            "format": "uuid",
            "title": "Channel Id",
            "type": "string"
          },
          "first_unread_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Unread At"
          },
          "last_read_seq": {
            "title": "Last Read Seq",
            "type": "integer"
          },
          "mention_count": {
            "title": "Mention Count",
            "type": "integer"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "channel_id",
          "last_read_seq",
          "unread_count",
          "mention_count"
        ],
        "title": "ChannelReadStateOut",
        "type": "object"
      },
      "ChannelUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[^\\s#@/]{1,80}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "posting_policy": {
            "anyOf": [
              {
                "enum": [
                  "everyone",
                  "owners"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Policy"
          },
          "purpose": {
            "anyOf": [
              {
                "maxLength": 250,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "times_owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Times Owner Id"
          },
          "topic": {
            "anyOf": [
              {
                "maxLength": 250,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "type": {
            "anyOf": [
              {
                "enum": [
                  "public",
                  "private"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "ChannelUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemberAdd": {
        "properties": {
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "MemberAdd",
        "type": "object"
      },
      "MemberOut": {
        "properties": {
          "joined_at": {
            "format": "date-time",
            "title": "Joined At",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "role",
          "joined_at"
        ],
        "title": "MemberOut",
        "type": "object"
      },
      "MemberRoleUpdate": {
        "additionalProperties": false,
        "description": "L4: PATCH /channels/{id}/members/{user_id}.",
        "properties": {
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "MemberRoleUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ChikuwaChat API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/channels": {
      "get": {
        "operationId": "list_channels_api_v1_channels_get",
        "parameters": [
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "public",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChannelOut"
                  },
                  "title": "Response List Channels Api V1 Channels Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Channels",
        "tags": [
          "channels"
        ]
      },
      "post": {
        "operationId": "create_channel_api_v1_channels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/v1/channels/read-all": {
      "post": {
        "description": "M12a 「すべて既読にする」: every channel I belong to is read to its end.",
        "operationId": "mark_all_read_api_v1_channels_read_all_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChannelReadStateOut"
                  },
                  "title": "Response Mark All Read Api V1 Channels Read All Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark All Read",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/v1/channels/{channel_id}": {
      "get": {
        "operationId": "get_channel_api_v1_channels__channel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Channel",
        "tags": [
          "channels"
        ]
      },
      "patch": {
        "operationId": "update_channel_api_v1_channels__channel_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/v1/channels/{channel_id}/archive": {
      "post": {
        "operationId": "archive_channel_api_v1_channels__channel_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Channel",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/channels/{channel_id}/join": {
      "post": {
        "operationId": "join_channel_api_v1_channels__channel_id__join_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Join Channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/v1/channels/{channel_id}/leave": {
      "post": {
        "operationId": "leave_channel_api_v1_channels__channel_id__leave_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Leave Channel",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/channels/{channel_id}/members": {
      "get": {
        "operationId": "list_members_api_v1_channels__channel_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberOut"
                  },
                  "title": "Response List Members Api V1 Channels  Channel Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Members",
        "tags": [
          "channels"
        ]
      },
      "post": {
        "operationId": "add_member_api_v1_channels__channel_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Member",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/v1/channels/{channel_id}/members/{user_id}": {
      "delete": {
        "operationId": "remove_member_api_v1_channels__channel_id__members__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Member",
        "tags": [
          "channels"
        ]
      },
      "patch": {
        "description": "L4: make a member an owner or a member again (owners and administrators).",
        "operationId": "update_member_api_v1_channels__channel_id__members__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Member",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/channels/{channel_id}/read": {
      "put": {
        "description": "Advance my read position (monotonic; SYNC_PROTOCOL.md §4.5 / §10).",
        "operationId": "mark_read_api_v1_channels__channel_id__read_put",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadMark"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Read",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/channels/{channel_id}/unarchive": {
      "post": {
        "description": "M13d: owner or administrator; posting works again afterwards.",
        "operationId": "unarchive_channel_api_v1_channels__channel_id__unarchive_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unarchive Channel",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/dms": {
      "post": {
        "operationId": "get_or_create_dm_api_v1_dms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DmCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Or Create Dm",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/times": {
      "post": {
        "description": "My times (M24): made on the first call (201), returned afterwards (200). The supervisors on\nthe lab roster join it (the lookup is injected by main.py: channels does not depend on lab).",
        "operationId": "ensure_times_api_v1_times_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ensure Times",
        "tags": [
          "channels"
        ]
      }
    },
...
  }
}